+++ src/Controller/AuthApiUserController.php
@@ -4,13 +4,17 @@ namespace CaliforniaMountainSnake\SimpleLaravelAuthSystem\Controller;
 
 use CaliforniaMountainSnake\SimpleLaravelAuthSystem\AccessUtils\AuthUserAccountTypeAccessUtils;
 use CaliforniaMountainSnake\SimpleLaravelAuthSystem\AccessUtils\AuthUserRoleAccessUtils;
-use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Controller\Utils\CreateUserEntityFromRequestApiToken;
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Authenticator\Interfaces\AuthenticatorInterface;
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Authenticator\Interfaces\HasUserInterface;
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Authenticator\Utils\HasAuthenticatorTrait;
+use CaliforniaMountainSnake\SimpleLaravelAuthSystem\Authenticator\Utils\HasUserTrait;
 use Illuminate\Contracts\Container\BindingResolutionException;
 
-abstract class AuthApiUserController extends AuthApiController
+abstract class AuthApiUserController extends AuthApiController implements HasUserInterface
 {
     use AuthApiUserControllerDependencies;
-    use CreateUserEntityFromRequestApiToken;
+    use HasAuthenticatorTrait;
+    use HasUserTrait;
     use AuthUserRoleAccessUtils;
     use AuthUserAccountTypeAccessUtils;
 
@@ -18,13 +22,16 @@ abstract class AuthApiUserController extends AuthApiController
 
     /**
      * AuthApiUserController constructor.
+     *
      * @throws BindingResolutionException
      */
     public function __construct()
     {
         parent::__construct();
         $this->initDependencies();
-        $this->initUserEntity();
+
+        $this->authenticator = app()->make(AuthenticatorInterface::class);
+        $this->userEntity = $this->authenticator->getUserEntity();
 
         if ($this->getUserEntity() !== null) {
             app()->setLocale((string)$this->getUserEntity()->getLanguage());